refactor : PointsGenerator 클래스 생성
