+++ src/main/java/coordinate/CoordinateCalculator.java
@@ -1,6 +1,7 @@
 package coordinate;
 
 import coordinate.domain.Point;
+import coordinate.domain.PointsGenerator;
 import coordinate.domain.Shape;
 import coordinate.domain.ShapeGenerator;
 import coordinate.view.InputView;
@@ -11,18 +12,29 @@ import java.util.List;
 public class CoordinateCalculator {
 
     public static void main(String[] args) {
-        Shape shape = createShape();
+        List<Point> points = createPoints();
+        Shape shape = createShape(points);
 
         OutputView.printArea(shape);
     }
 
-    private static Shape createShape() {
+    private static List<Point> createPoints() {
         try {
-            List<Point> points = ShapeGenerator.generatePoints(InputView.inputCoordinate());
+            return PointsGenerator.generatePoints(InputView.inputCoordinate());
+        } catch (IllegalArgumentException e) {
+            System.out.println(e.getMessage());
+            return createPoints();
+        }
+    }
+
+    private static Shape createShape(List<Point> points) {
+        try {
+
             return ShapeGenerator.create(points);
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
-            return createShape();
+            points = createPoints();
+            return createShape(points);
         }
     }
 }
